+++ release_notes.md
@@ -1,3 +1,11 @@
+## Unreleased
+- Adding Snapshot class to register images as notes in Alyx
+- Ephys pipeline: RawEphysQC: outputs channels labels as part of pipeline
+- Ephys pipeline: spike sorting
+    - dsp.voltage: destripe detects bad channels and interpolate them
+    - synchronisation 3B: gives a 10% leeway to throw exception
+    - spikes sorting task outputs the RMS plot
+
 ## Release Notes 2.4
 ### Release Notes 2.4.0
 - Setting tasks to Waiting if they encountered lock (status -2)